+++ tests/fork/cron-sigaction-query-failure.sh
@@ -1,7 +1,7 @@
 . "$(dirname "$0")"/../util/common.sh
 
 TEST_LD_PRELOAD=(test-sigaction-query-failure test-fake-getpid test-fake-getuid)
-run_test "$RUN"
+run_test "-q" "$RUN"
 RET=$?
 
-variables_must_eq RET $EX_UNAVAILABLE
+variables_must_eq RET $EXIT_SUCCESS