+++ tests/fork/sigaction-failure.sh
@@ -1,6 +1,6 @@
 . "$(dirname "$0")"/../util/common.sh
 
-TEST_LD_PRELOAD=(test-sigaction-set-restart-failure test-fake-getpid test-fake-getuid)
+TEST_LD_PRELOAD=(test-sigaction-failure test-fake-getpid test-fake-getuid)
 run_test "$RUN"
 RET=$?